rename steps subproject to behaviour_builder
