+++ settings.gradle
@@ -8,6 +8,6 @@ pluginManagement {
 include ":tgbotapi.core"
 include ":tgbotapi.extensions.api"
 include ":tgbotapi.extensions.utils"
-include ":tgbotapi.extensions.steps"
+include ":tgbotapi.extensions.behaviour_builder"
 include ":tgbotapi"
 include ":docs"
