+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/expectations/WaitContent.kt
@@ -1,8 +1,8 @@
 @file:Suppress("unused")
 
-package dev.inmo.tgbotapi.extensions.steps.expectations
+package dev.inmo.tgbotapi.extensions.behaviour_builder.expectations
 
-import dev.inmo.tgbotapi.extensions.steps.BehaviourContext
+import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
 import dev.inmo.tgbotapi.extensions.utils.asContentMessage
 import dev.inmo.tgbotapi.extensions.utils.asMessageUpdate
 import dev.inmo.tgbotapi.requests.abstracts.Request
