+++ tgbotapi/build.gradle
@@ -42,7 +42,7 @@ kotlin {
                 api project(":tgbotapi.core")
                 api project(":tgbotapi.extensions.api")
                 api project(":tgbotapi.extensions.utils")
-                api project(":tgbotapi.extensions.steps")
+                api project(":tgbotapi.extensions.behaviour_builder")
             }
         }
     }
